Refactor Event-models
