+++ apps/events/models/Attendance.py
@@ -1,7 +1,3 @@
-# -*- coding: utf-8 -*-
-
-import logging
-from collections import OrderedDict
 from datetime import datetime, timedelta
 from random import choice as random_choice
 from typing import List
@@ -9,528 +5,23 @@ from typing import List
 from django.conf import settings
 from django.contrib.auth.models import Group
 from django.contrib.contenttypes.fields import GenericRelation
-from django.core import validators
-from django.core.exceptions import (
-    MultipleObjectsReturned,
-    ObjectDoesNotExist,
-    ValidationError,
-)
+from django.core.exceptions import MultipleObjectsReturned, ObjectDoesNotExist
 from django.core.mail import EmailMessage
 from django.db import models
-from django.db.models import SET_NULL, Case, Q, Value, When
-from django.template.defaultfilters import slugify
 from django.template.loader import render_to_string
-from django.urls import reverse
 from django.utils import timezone
 from django.utils.translation import gettext as _
 from guardian.shortcuts import assign_perm
-from unidecode import unidecode
 
-from apps.authentication.constants import FieldOfStudyType
 from apps.authentication.models import OnlineGroup
 from apps.companyprofile.models import Company
-from apps.events.constants import EventType
-from apps.feedback.models import FeedbackRelation
-from apps.gallery.models import ResponsiveImage
 from apps.marks.models import get_expiration_date
 from apps.payment import status as payment_status
 from apps.payment.mixins import PaymentMixin
 
-logger = logging.getLogger(__name__)
-
-User = settings.AUTH_USER_MODEL
-
-# Managers
-
-
-class EventOrderedByRegistration(models.Manager):
-    """
-    Order events by registration start if registration start is within 7 days of today.
-    """
-
-    def get_queryset(self):
-        now = timezone.now()
-        DELTA_FUTURE_SETTING = settings.OW4_SETTINGS.get("events").get(
-            "FEATURED_DAYS_FUTURE"
-        )
-        DELTA_PAST_SETTING = settings.OW4_SETTINGS.get("events").get(
-            "FEATURED_DAYS_PAST"
-        )
-        DAYS_BACK_DELTA = timezone.now() - timedelta(days=DELTA_PAST_SETTING)
-        DAYS_FORWARD_DELTA = timezone.now() + timedelta(days=DELTA_FUTURE_SETTING)
-
-        return (
-            super(EventOrderedByRegistration, self)
-            .get_queryset()
-            .annotate(
-                registration_filtered=Case(
-                    When(
-                        Q(event_end__gte=now)
-                        & Q(attendance_event__registration_start__gte=DAYS_BACK_DELTA)
-                        & Q(
-                            attendance_event__registration_start__lte=DAYS_FORWARD_DELTA
-                        ),
-                        then="attendance_event__registration_start",
-                    ),
-                    default="event_end",
-                    output_field=models.DateTimeField(),
-                )
-            )
-            .annotate(
-                is_today=Case(
-                    When(event_end__date=now.date(), then=Value(1)),
-                    default=Value(0),
-                    output_field=models.IntegerField(),
-                )
-            )
-            .order_by("-is_today", "registration_filtered")
-        )
-
-    def get_queryset_for_user(self, user: User):
-        """
-        :return: Queryset filtered by these requirements:
-            event is visible AND (event has NO group restriction OR user having access to restricted event)
-            OR the user is attending the event themselves
-        """
-        group_restriction_query = Q(group_restriction__isnull=True) | Q(
-            group_restriction__groups__in=user.groups.all()
-        )
-        is_attending_query = (
-            (
-                Q(attendance_event__isnull=False)
-                & Q(attendance_event__attendees__user=user)
-            )
-            if not user.is_anonymous
-            else Q()
-        )
-        is_visible_query = Q(visible=True)
-        return (
-            self.get_queryset()
-            .filter(group_restriction_query & is_visible_query | is_attending_query)
-            .distinct()
-        )
-
-
-class Event(models.Model):
-    """
-    Base class for Event-objects.
-    """
-
-    IMAGE_FOLDER = "images/events"
-    IMAGE_EXTENSIONS = [".jpg", ".jpeg", ".gif", ".png", ".tif", ".tiff"]
-
-    # Managers
-    objects = models.Manager()
-    by_registration = EventOrderedByRegistration()
-
-    author = models.ForeignKey(
-        User,
-        related_name="created_events",
-        verbose_name=_("Oppretter"),
-        null=True,
-        blank=True,
-        on_delete=models.CASCADE,
-    )
-    title = models.CharField(_("tittel"), max_length=60)
-    """Event title"""
-    event_start = models.DateTimeField(_("start-dato"))
-    """Datetime for event start"""
-    event_end = models.DateTimeField(_("slutt-dato"))
-    """Datetime for event end"""
-    location = models.CharField(_("lokasjon"), max_length=100)
-    """Event location"""
-    ingress_short = models.CharField(
-        _("kort ingress"),
-        max_length=150,
-        validators=[validators.MinLengthValidator(25)],
-        help_text="En kort ingress som blir vist på forsiden",
-    )
-    """Short ingress used on the frontpage"""
-    ingress = models.TextField(
-        _("ingress"),
-        validators=[validators.MinLengthValidator(25)],
-        help_text="En ingress som blir vist før beskrivelsen.",
-    )
-    """Ingress used in archive and details page"""
-    description = models.TextField(
-        _("beskrivelse"), validators=[validators.MinLengthValidator(45)]
-    )
-    """Event description shown on details page"""
-    image = models.ForeignKey(
-        ResponsiveImage,
-        related_name="events",
-        blank=True,
-        null=True,
-        on_delete=SET_NULL,
-    )
-    """Event image"""
-    event_type = models.SmallIntegerField(
-        _("type"), choices=EventType.ALL_CHOICES, null=False
-    )
-    """Event type. Used mainly for filtering"""
-    organizer = models.ForeignKey(
-        Group, verbose_name=_("arrangør"), blank=True, null=True, on_delete=SET_NULL
-    )
-    """Committee responsible for organizing the event"""
-    visible = models.BooleanField(
-        _("Vis arrangementet utenfor Dashboard og Adminpanelet"),
-        default=True,
-        help_text=_("Denne brukes for å skjule eksisterende arrangementer."),
-    )
-    companies = models.ManyToManyField(
-        to=Company,
-        verbose_name=_("Bedrifter"),
-        related_name="events",
-        through="CompanyEvent",
-    )
-
-    feedback = GenericRelation(FeedbackRelation)
-
-    def is_attendance_event(self):
-        """ Returns true if the event is an attendance event """
-        return hasattr(self, "attendance_event")
-
-    # TODO move payment and feedback stuff to attendance event when dasboard is done
-
-    def feedback_users(self):
-        if self.is_attendance_event():
-            return [
-                a.user for a in self.attendance_event.attendees.filter(attended=True)
-            ]
-        return []
-
-    def feedback_date(self):
-        return self.event_end
-
-    def feedback_title(self):
-        return self.title
-
-    def feedback_info(self):
-        info = OrderedDict()
-        if self.is_attendance_event():
-            info[_("Påmeldte")] = self.attendance_event.number_of_attendees
-            info[_("Oppmøtte")] = self.attendance_event.number_of_attendees - len(
-                self.attendance_event.not_attended()
-            )
-            info[_("Venteliste")] = self.attendance_event.number_on_waitlist
-
-        return info
-
-    @property
-    def images(self):
-        images = ResponsiveImage.objects.none()
-        if self.image:
-            images |= ResponsiveImage.objects.filter(pk=self.image.id)
-        company_image_ids = self.companies.values_list("image")
-        images |= ResponsiveImage.objects.filter(pk__in=company_image_ids)
-        return images.distinct()
-
-    @property
-    def company_event(self):
-        return CompanyEvent.objects.filter(event=self)
-
-    @property
-    def organizer_name(self):
-        return self.organizer.name
-
-    def feedback_mail(self):
-        """
-        Get feedback mail from organizer Online group.
-        Use default from email if the organizer group does not have configured an email.
-        """
-        if self.organizer:
-            organizer_group = OnlineGroup.objects.filter(pk=self.organizer.id).first()
-            if organizer_group and organizer_group.email:
-                return organizer_group.email
-
-        return settings.DEFAULT_FROM_EMAIL
-
-    def can_display(self, user):
-        restriction = GroupRestriction.objects.filter(event=self).first()
-        if (
-            not user.is_anonymous
-            and self.is_attendance_event()
-            and self.attendance_event.is_attendee(user)
-        ):
-            return True
-        if not self.visible:
-            return False
-        if not restriction:
-            return True
-        if not user:
-            return False
-        return restriction.has_access(user)
-
-    @property
-    def slug(self):
-        return slugify(unidecode(self.title))
-
-    def get_absolute_url(self):
-        return reverse(
-            "events_details", kwargs={"event_id": self.id, "event_slug": self.slug}
-        )
-
-    def clean(self):
-        if not self.organizer:
-            raise ValidationError({"organizer": ["Arrangementet krever en arrangør."]})
-
-    def save(
-        self, force_insert=False, force_update=False, using=None, update_fields=None
-    ):
-        super().save(
-            force_insert=force_insert,
-            force_update=force_update,
-            using=using,
-            update_fields=update_fields,
-        )
-
-        if self.organizer:
-            assign_perm("events.change_event", self.organizer, obj=self)
-            assign_perm("events.delete_event", self.organizer, obj=self)
-
-    def __str__(self):
-        return self.title
-
-    class Meta:
-        verbose_name = _("arrangement")
-        verbose_name_plural = _("arrangementer")
-        permissions = (("view_event", "View Event"),)
-        default_permissions = ("add", "change", "delete")
-
-
-"""
- BEGIN ACCESS RESTRICTION --------------------------------------------------------------------------
-"""
-
-
-class Rule(models.Model):
-    """
-    Super class for a rule object
-    """
-
-    success_code = 213
-    not_opened_code = 402
-    delayed_signup_code = 423
-    delayed_singup_message = _("Du har utsatt påmelding")
-    not_satisfied_code = 413
-    not_satisfied_message = _("Du har ikke tilgang til å melde deg på arrangementet")
-
-    offset = models.PositiveSmallIntegerField(
-        _("utsettelse"), help_text=_("utsettelse oppgis i timer"), default=0
-    )
-
-    def get_offset_time(self, time: timezone.datetime):
-        if type(time) is not datetime:
-            raise TypeError("time must be a datetime, not %s" % type(time))
-        else:
-            return time + timedelta(hours=self.offset)
-
-    def satisfies_constraint(self, user: User) -> bool:
-        return True
-
-    def satisfied(self, user: User, registration_start: timezone.datetime):
-        """ Override method """
-
-        if self.satisfies_constraint(user):
-            offset_datetime = self.get_offset_time(registration_start)
-            # If the offset is in the past, it means you can attend even with the offset
-            if offset_datetime < timezone.now():
-                return {
-                    "status": True,
-                    "message": None,
-                    "status_code": self.success_code,
-                }
-            # If there is no offset, the signup just hasn't started yet
-            elif self.offset == 0:
-                return {
-                    "status": False,
-                    "message": _("Påmeldingen har ikke åpnet enda."),
-                    "status_code": self.not_opened_code,
-                }
-            # In the last case there is a delayed signup
-            else:
-                return {
-                    "status": False,
-                    "message": self.delayed_singup_message,
-                    "offset": offset_datetime,
-                    "status_code": self.delayed_signup_code,
-                }
-        return {
-            "status": False,
-            "message": self.not_satisfied_message,
-            "status_code": self.not_satisfied_code,
-        }
-
-    def __str__(self):
-        return "Rule"
-
-    class Meta:
-        permissions = (("view_rule", "View Rule"),)
-        default_permissions = ("add", "change", "delete")
-
-
-class FieldOfStudyRule(Rule):
-    success_code = 210
-    delayed_signup_code = 420
-    delayed_singup_message = _("Din studieretning har utsatt påmelding.")
-    not_satisfied_code = 410
-    not_satisfied_message = _(
-        "Din studieretning er en annen enn de som har tilgang til dette arrangementet."
-    )
-
-    field_of_study = models.SmallIntegerField(
-        _("studieretning"), choices=FieldOfStudyType.ALL_CHOICES
-    )
-
-    def satisfies_constraint(self, user: User) -> bool:
-        return self.field_of_study == user.field_of_study
-
-    def __str__(self):
-        if self.offset > 0:
-            time_unit = _("timer") if self.offset > 1 else _("time")
-            return _("%s etter %d %s") % (
-                str(self.get_field_of_study_display()),
-                self.offset,
-                time_unit,
-            )
-        return str(self.get_field_of_study_display())
-
-    class Meta:
-        permissions = (("view_fieldofstudyrule", "View FieldOfStudyRule"),)
-        default_permissions = ("add", "change", "delete")
-
-
-class GradeRule(Rule):
-    success_code = 211
-    delayed_signup_code = 421
-    delayed_singup_message = _("Ditt klassetrinn har utsatt påmelding.")
-    not_satisfied_code = 411
-    not_satisfied_message = _(
-        "Ditt klassetrinn har ikke tilgang til dette arrangementet.."
-    )
-
-    grade = models.SmallIntegerField(_("klassetrinn"), null=False)
-
-    def satisfies_constraint(self, user: User) -> bool:
-        return self.grade == user.year
-
-    def __str__(self):
-        if self.offset > 0:
-            time_unit = _("timer") if self.offset > 1 else _("time")
-            return _("%s. klasse etter %d %s") % (self.grade, self.offset, time_unit)
-        return _("%s. klasse") % self.grade
-
-    class Meta:
-        permissions = (("view_graderule", "View GradeRule"),)
-        default_permissions = ("add", "change", "delete")
-
-
-class UserGroupRule(Rule):
-    success_code = 212
-    delayed_signup_code = 422
-    delayed_singup_message = _("Brukergruppene dine har utsatt påmelding")
-    not_satisfied_code = 412
-    not_satisfied_message = _(
-        "Din brukergruppe har ikke tilgang til dette arrangementet."
-    )
-
-    group = models.ForeignKey(Group, blank=False, null=False, on_delete=models.CASCADE)
-
-    def satisfies_constraint(self, user: User) -> bool:
-        return self.group in user.groups.all()
-
-    def __str__(self):
-        if self.offset > 0:
-            time_unit = _("timer") if self.offset > 1 else _("time")
-            return _("%s etter %d %s") % (str(self.group), self.offset, time_unit)
-        return str(self.group)
-
-    class Meta:
-        permissions = (("view_usergrouprule", "View UserGroupRule"),)
-        default_permissions = ("add", "change", "delete")
-
-
-class RuleBundle(models.Model):
-    """
-    Access restriction rule object
-    """
-
-    description = models.CharField(
-        _("beskrivelse"), max_length=100, blank=True, null=True
-    )
-    field_of_study_rules = models.ManyToManyField(FieldOfStudyRule, blank=True)
-    grade_rules = models.ManyToManyField(GradeRule, blank=True)
-    user_group_rules = models.ManyToManyField(UserGroupRule, blank=True)
-
-    def get_all_rules(self):
-        rules: List[Rule] = []
-        rules.extend(self.field_of_study_rules.all())
-        rules.extend(self.grade_rules.all())
-        rules.extend(self.user_group_rules.all())
-        return rules
-
-    @property
-    def rule_strings(self):
-        return [str(rule) for rule in self.get_all_rules()]
-
-    def satisfied(self, user, registration_start):
-
-        errors = []
-
-        for rule in self.get_all_rules():
-            response = rule.satisfied(user, registration_start)
-            if response["status"]:
-                return [response]
-            else:
-                errors.append(response)
-
-        return errors
-
-    def get_minimum_offset_for_user(self, user: User):
-        offsets = [
-            rule.offset
-            for rule in self.get_all_rules()
-            if rule.satisfies_constraint(user)
-        ]
-        if len(offsets) == 0:
-            return timezone.timedelta(hours=0)
-        offsets.sort()
-        return timezone.timedelta(hours=offsets[0])
-
-    def __str__(self):
-        if self.description:
-            return self.description
-        elif self.rule_strings:
-            return ", ".join(self.rule_strings)
-        else:
-            return _("Tom rule bundle.")
-
-    class Meta:
-        permissions = (("view_rulebundle", "View RuleBundle"),)
-        default_permissions = ("add", "change", "delete")
-
-
-"""
- END ACCESS RESTRICTION --------------------------------------------------------------------------
-"""
-
-
-class Extras(models.Model):
-    """
-    Choices for events
-    """
-
-    choice = models.CharField("valg", max_length=69)
-    note = models.CharField("notat", max_length=200, blank=True, null=True)
-
-    def __str__(self):
-        return self.choice
-
-    class Meta:
-        verbose_name = _("ekstra valg")
-        verbose_name_plural = _("ekstra valg")
-        ordering = ["choice"]
-        default_permissions = ("add", "change", "delete")
+from .AccessRestriction import RuleBundle
+from .Event import Event, User, logger
+from .Extras import Extras
 
 
 class AttendanceEvent(PaymentMixin, models.Model):
