+++ apps/events/models/__init__.py
@@ -1,0 +1,17 @@
+from .AccessRestriction import (
+    FieldOfStudyRule,
+    GradeRule,
+    Rule,
+    RuleBundle,
+    UserGroupRule,
+)
+from .Attendance import (
+    AttendanceEvent,
+    Attendee,
+    CompanyEvent,
+    GroupRestriction,
+    Reservation,
+    Reservee,
+)
+from .Event import Event, EventOrderedByRegistration
+from .Extras import Extras
